govtuner: Smarter cluster detection system  [TEST ONLY]
We put it here for testing, if it works we'll put it up to profiles, but let's put it here because it's safer.

Background:
Cluster-equipped SoC also face a problem when there's 6 or more cores. On quad / 4 cores, the cores is split evenly, two big and two small. However, when it reaches more than four, say for example, 6, the cores aren't being split evenly, it'll be splitted to 4 small and two big. In 8 cores, this situation is better because it is mostly split by four big and four small.

How it works:
This system detects the maximum frequency of CPU0, CPU2, and CPU4 (first core, third core, and fifth core respectively), because the big and small cores cannot have the same maximum (the steps are different), the cores will always be different, unless you are idiotic enough to set them to the absolute minimum step. We add the fifth core to avoid the stupidity of the 6 core device (where the fourth core can be the same as third core).

Additionally, we use CPU4 instead of CPU3 because CPU4 doesn't exist in quad core cluster SoC, this increases the chance of correct detection since there's less variable to work on if CPU4 doesn't exist.

v1 -> Initially merge this