Add buildscript for Windows
This version is largely untested, it's a quick port over the Linux
version.

Provided with this commit is the "zip" binary for Windows, this is due
to the fact that most zip installation for Windows are either localised
(only for a certain program / app), to workaround the issue, we provide
our own copy of the "zip" binary and its dependency.

The copy of the binary is built by GnuWin / gnuwin32 @ sourceforge.